requirements.txt
# VerifDoc - Dependencies
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Image processing & forensics
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
PyWavelets>=1.5.0

# PDF handling
PyMuPDF>=1.23.0

# OCR
easyocr>=1.7.1

# Dashboard
streamlit>=1.28.0

# Utilities
python-dateutil>=2.8.0
